+++ src/fields/Entries.php
@@ -8,6 +8,7 @@
 namespace craft\fields;
 
 use Craft;
+use craft\base\GqlSchemaAwareFieldInterface;
 use craft\db\Table as DbTable;
 use craft\elements\db\EntryQuery;
 use craft\elements\Entry;
@@ -24,7 +25,7 @@ use GraphQL\Type\Definition\Type;
  * @author Pixel & Tonic, Inc. <support@pixelandtonic.com>
  * @since 3.0.0
  */
-class Entries extends BaseRelationField
+class Entries extends BaseRelationField implements GqlSchemaAwareFieldInterface
 {
     /**
      * @inheritdoc
@@ -93,5 +94,14 @@ class Entries extends BaseRelationField
             'typeId' => array_values($entryTypeIds),
             'sectionId' => array_values($sectionIds)
         ];
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function getExistsForCurrentGqlSchema(): bool
+    {
+        $allowedEntities = Gql::extractAllowedEntitiesFromSchema();
+        return !empty($allowedEntities['sections']) && !empty($allowedEntities['entrytypes']);
     }
 }